feature(TSK-012): sass variables
