+++ src/components/Categories/CategoryList.module.scss
@@ -1,0 +1,14 @@
+.categoryList {
+  display: flex;
+  flex-direction: row;
+  border: 1px solid #e1e5e9;
+  border-radius: 12px;
+  padding: 1.5rem;
+  margin-top: 2rem;
+  background: white;
+
+  .categoryTab {
+    padding: 0;
+    margin: 0;
+  }
+}
