+++ src/components/Categories/CategoryTab.tsx
@@ -1,6 +1,8 @@
 import React from 'react'
 import { useDispatch } from "react-redux";
 
+import styles from './CategoryList.module.scss'
+
 import { IItems } from '@/features/lobby/typings';
 import { changeFilter } from '@/features/lobby/actions';
 
@@ -10,22 +12,22 @@ interface IOwnProps {
 
 export const CategoryTab = ( props: IOwnProps ) => {
   const { category } = props;
+  const { id = -1, categoryFilter = '', name: { en = ''} } = category;
   const dispatch = useDispatch();
 
-  
-  const handleOnClick = (categoryFilter: string | undefined) => {
-    if (!categoryFilter) {
+  const handleOnClick = () => {
+    if (categoryFilter.length === 0) {
       return false;
     }
     // TODO - send Category in games API properly as to Apply filter from BE side
     dispatch(changeFilter(categoryFilter));
   };
 
-  return <li key={category.id}>
+  return <div key={id} className={styles.categoryTab} >
     {
-      category.categoryFilter && category.categoryFilter.length > 0
-        ? <button onClick={() => handleOnClick(category.categoryFilter)}><span>{category.name.en}</span></button>
-        : <span>{category.name.en}</span>
+      categoryFilter.length > 0
+        ? <button onClick={handleOnClick}><span>{en}</span></button>
+        : <span>{en}</span>
     }
-  </li>;
+  </div>;
 }
