+++ src/features/game/constants.ts
@@ -1,2 +1,4 @@
 export const NEW_GAME = "NEW_GAME"
 export const POPULARITY_UPDATE = "POPULARITY_UPDATE"
+
+export const API_BASE = 'https://casino.api.pikakasino.com/v1/pika';
