+++ src/features/lobby/constants.ts
@@ -11,3 +11,5 @@ export const initialState: ILobbyState = {
     category: null,
     searchQuery: { name: '', pageSize: 100 },
 };
+
+export const API_BASE = 'https://casino.api.pikakasino.com/v1/pika';
