+++ src/styles/globals.scss
@@ -1,10 +1,36 @@
-$primary: #f5f5f5;
+@import './variables.scss';
+
+// Browser variables
+:root {
+  --base-unit: 1rem;
+  --unit: var( --base-unit );
+
+  --background-color: #{$secondary};
+  --text-color: #{$primary};
+  --text-color-secondary: #{$blue-lynch};
+  --primary-glow: #{$orange-flamingo};
+  --secondary-glow: #{$blue-clear};
+  --callout-rgb: #{$secondary};
+  --callout-border-rgb: #{$blue-lynch};
+  --card-rgb: #{$secondary};
+  --card-border-rgb: #{$blue-lynch};
+  --background-end-rgb: #{$blue-clear};
+
+  --max-width: 1200px;
+  --border-radius: units(1);
+  --font-mono: 'Fira Code', monospace;
+
+  @media ( max-width: 470px ) { // Iphones and mobiles smaller then 480px
+    --base-unit: 0.8rem;
+  }
+}
 
 html, body {
   margin: 0;
   padding: 0;
+  font-size: calc( 12 / 16 * var( --unit , 1rem ) );
   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Oxygen, Ubuntu, Cantarell, "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-  background-color: $primary;
+  background-color: $secondary;
   color: #333;
 }
 
@@ -24,6 +50,6 @@ button {
 .grid {
   display: grid;
   grid-template-columns: repeat(auto-fill, minmax(220px, 1fr));
-  gap: 16px;
-  padding: 16px;
+  gap: units(16);
+  padding: units(16);
 }
